Switch implementation to use a Definition model
