+++ core/resource_definitions.go
@@ -16,6 +16,7 @@ package core
 
 import (
 	"context"
+	"fmt"
 
 	"github.com/sacloud/libsacloud/v2/sacloud"
 )
@@ -39,7 +40,7 @@ func (r *ResourceDefinitions) Validate(ctx context.Context, apiClient sacloud.AP
 	return errors
 }
 
-type ResourceDefWalkFunc func(definitions ResourceDefinition) error
+type ResourceDefWalkFunc func(def ResourceDefinition) error
 
 // Walk 各リソースに対し順次fnを適用する
 //
@@ -83,6 +84,105 @@ func (r *ResourceDefinitions) walk(targets ResourceDefinitions, fn ResourceDefWa
 			if err := r.walk(child.Children(), fn); err != nil {
 				return err
 			}
+		}
+	}
+	return nil
+}
+
+func (r *ResourceDefinitions) HandleAll(ctx *RequestContext, apiClient sacloud.APICaller, handlers Handlers) error {
+	return r.handleAll(ctx, apiClient, handlers, nil, *r)
+}
+
+func (r *ResourceDefinitions) handleAll(ctx *RequestContext, apiClient sacloud.APICaller, handlers Handlers, parentResource Resource2, defs ResourceDefinitions) error {
+	for _, def := range defs {
+		resources, err := def.Compute(ctx, apiClient)
+		if err != nil {
+			return err
+		}
+		// 子リソースが定義されているリソースは複数ヒット時にはエラーとする
+		children := def.Children()
+
+		if len(children) > 0 && len(resources) > 1 {
+			return fmt.Errorf("A resource definition with children cannot return multiple resources: definition: %#v, returned: %#v", def, resources)
+		}
+
+		for _, resource := range resources {
+			if parentResource != nil {
+				resource.SetParent(parentResource)
+			}
+			if len(children) > 0 {
+				if err := r.handleAll(ctx, apiClient, handlers, resource, children); err != nil {
+					return err
+				}
+			}
+
+			if err := r.handleResource(ctx, handlers, resource); err != nil {
+				return err
+			}
+		}
+	}
+	return nil
+}
+
+func (r *ResourceDefinitions) handleResource(parentCtx *RequestContext, handlers Handlers, resource Resource2) error {
+	computed, err := resource.Compute(parentCtx, false)
+	if err != nil {
+		return err
+	}
+
+	zone := computed.Zone()
+	if zone == "" {
+		zone = "global"
+	}
+	handlingCtx := NewHandlingContext(parentCtx, computed).WithLogger("type", computed.Type(), "zone", zone, "id", computed.ID())
+
+	// preHandle
+	if err := r.handleAllByFunc(computed, handlers, func(h *Handler, c Computed) error {
+		ctx := handlingCtx.WithLogger("step", "PreHandle", "handler", h.Name)
+		if h.BuiltinHandler != nil {
+			h.BuiltinHandler.SetLogger(ctx.Logger())
+		}
+		return h.PreHandle(ctx, c)
+	}); err != nil {
+		return err
+	}
+
+	// handle
+	if err := r.handleAllByFunc(computed, handlers, func(h *Handler, c Computed) error {
+		ctx := handlingCtx.WithLogger("step", "Handle", "handler", h.Name)
+		if h.BuiltinHandler != nil {
+			h.BuiltinHandler.SetLogger(ctx.Logger())
+		}
+		return h.Handle(ctx, c)
+	}); err != nil {
+		return err
+	}
+
+	// refresh
+	refreshed, err := resource.Compute(handlingCtx.RequestContext, true)
+	if err != nil {
+		return err
+	}
+	computed = refreshed
+
+	// postHandle
+	if err := r.handleAllByFunc(computed, handlers, func(h *Handler, c Computed) error {
+		ctx := handlingCtx.WithLogger("step", "PostHandle", "handler", h.Name)
+		if h.BuiltinHandler != nil {
+			h.BuiltinHandler.SetLogger(ctx.Logger())
+		}
+		return h.PostHandle(ctx, c)
+	}); err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func (r *ResourceDefinitions) handleAllByFunc(computed Computed, handlers Handlers, fn func(*Handler, Computed) error) error {
+	for _, handler := range handlers {
+		if err := fn(handler, computed); err != nil {
+			return err
 		}
 	}
 	return nil
