+++ core/resource_plans_test.go
