--- core/resource_server_group.go
@@ -1,46 +1,0 @@
-// Copyright 2021 The sacloud Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package core
-
-import (
-	"context"
-
-	"github.com/sacloud/libsacloud/v2/sacloud"
-)
-
-type ServerGroup struct {
-	*ResourceBase `yaml:",inline"`
-	wrapper       Resource
-}
-
-func (s *ServerGroup) Validate(ctx context.Context, apiClient sacloud.APICaller) []error {
-	// TODO 実装
-	return nil
-}
-
-func (s *ServerGroup) Compute(ctx *RequestContext, apiClient sacloud.APICaller) (Computed, error) {
-	// TODO 実装
-	return nil, nil
-}
-
-// Parent ChildResourceインターフェースの実装
-func (s *ServerGroup) Parent() Resource {
-	return s.wrapper
-}
-
-// SetParent ChildResourceインターフェースの実装
-func (s *ServerGroup) SetParent(parent Resource) {
-	s.wrapper = parent
-}
